+++ src/main/java/markets/alpaca/client/openapi/broker/model/JournalStatusEventV2.java
@@ -510,8 +510,6 @@ public class JournalStatusEventV2 implements Serializable {
       }
       // validate the required field `status_from`
       JournalStatusFrom.validateJsonElement(jsonObj.get("status_from"));
-      // validate the required field `status_from`
-      JournalStatusFrom.validateJsonElement(jsonObj.get("status_from"));
       // validate the required field `status_to`
       JournalStatus.validateJsonElement(jsonObj.get("status_to"));
   }